+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+0.15.0 - (2025-11-30)
+------------------
+
+* Added `--resize-inputs` option to force resizing of inputs in cases where multiple videos are
+input with different resolutions.
+* Added `--output-resolution` and supporting click types to set the output resolution explicitly. 
+
+
 0.14.1 - (2025-11-30)
 ------------------